Renaming logstah_logging.rb -> logstash_logging.rb
