+++ lib/sidekiq/logstash.rb
@@ -1,6 +1,6 @@
 require 'sidekiq/logstash/configuration'
 require 'sidekiq/logstash/version'
-require 'sidekiq/middleware/server/logstah_logging'
+require 'sidekiq/middleware/server/logstash_logging'
 require 'sidekiq/logging/logstash_formatter'
 require 'sidekiq/logging/argument_filter'
 require 'sidekiq/logstash_job_logger'
