+++ lib/sidekiq/middleware/server/logstash_logging.rb
